requirements.txt
# Usage: pip install -r requirements.txt
#python version = 3.7.11
#conda create --name face_detect python==3.7.11

# Base ----------------------------------------
matplotlib>=3.4.2
numpy>=1.21.6
opencv-python>=4.1.1
Pillow>=6.1.0
PyYAML>=5.4.1
requests>=2.28.1
scipy>=1.6.3
torch>=1.7.0
torchvision>=0.13.1
tqdm>=4.64.0
tensorflow>=2.9.1
scikit-learn==1.0.2
joblib=1.3.2

# Logging -------------------------------------
tensorboard>=2.9.1

# Plotting ------------------------------------
pandas>=1.3.5
seaborn>=0.11.2

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation

# For API
streamlit
